Move Timer class to util module
